+++ CHANGELOG.md
@@ -7,8 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Add additional callback in AudioVideoObserver to indicate video downlink pressure
+
 ### Changed
 - Prevent prebuild from increase patch number when publishing to NPM
+
+
 
 ## [1.2.1] - 2020-03-20